requirements.txt
# Core Dependencies
groq
google-generativeai
pydantic
pydantic-settings
python-dotenv

# Web Framework
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Processing

# RAG & Embeddings (Optional - for advanced RAG)
# sentence-transformers==2.2.2
# chromadb==0.4.22

# Utilities
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
